rfe15951: 1.3.68: computed-content for do-http-request
Allow content in a call to do-http-request to be generated
at HTTP request time and sent to the web server.

%Testing: cd aserve; make test

<release-note>
rfe15951

do-http-request's content argument can now be a subclass of
computed-content.  In this case the content will be generated
at HTTP request time and sent to the web server.
</release-note>

<documentation>
rfe15951
 aserve.html updated with the new documentation

</documentation>

Change-Id: Ib93291db7bcc027e4fd0c4e0e736ade54484ff0a
Reviewed-on: https://gerrit.franz.com:9080/12290
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Martin Mikelsons <mm@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>