rfe16018: 1.3.69: automatic caching in the client
An HTTP server returns headers that specify how long the response can be cached.
However the client may wish to cache pages despite the server not
specifying a cache time (the client is willing to gamble that it knows
better than the server how long a response is valid).
With this change the client can now cause caching to occur.

%Testing: make test

<documentation>
rfe16018

The aserve.html document has been updated to include all the
client-cache documentation including the following changes
done in this submission:

Add two initargs to a make-instance of net.aserve.client:client-cache

:auto-cache-codes - a list of HTTP response codes that should be cached
:auto-cache-seconds - the number of seconds an instance will be cached
    if no cache information is given in the response and if the response
    code is one of the codes in the auto-cache-codes list.

</documentation>

Change-Id: I604a13d310d12ef6b8d31ef2924072bc201c50cb
Reviewed-on: https://gerrit.franz.com:9080/12584
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Reviewed-by: Martin Mikelsons <mm@franz.com>
Tested-by: Kevin Layer <layer@franz.com>