bug25698: make stress-aserve is broken on 10.1
%Testing: make stress-aserve

The test test-get-request-body-inc was writing and deleting a file
with a constant name.  This created conflicts in stress test where
several threads were writing and deleting thew same file.

Modify the file name with the test instance index, and add a
comment at the top of the file for future test writers.

Change-Id: Ib4dd89514d5481880837e3741d7d8e362bfa4aa6
Reviewed-on: https://gerrit.franz.com:9080/12818
Reviewed-by: Martin Mikelsons <mm@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>