rfe16148: 1.3.73: allow free worker wait timeout configuration
When accepting connections, function NET.ASERVE:HTTP-ACCEPT-THREAD
looks for a free worker thread with a hardcoded timeout of 1 seconds
for a few times before creating a new one.

This change implements means to control that timeout via a
*HTTP-FREE-WORKER-TIMEOUT* dynamic variable to be able to adapt to
situations when we need to quickly create a lot of worker threads.

This also simplifies worker allocation loop to waiting for
*HTTP-FREE-WORKER-TIMEOUT* seconds only once before creating a new
thread, instead of up to 5 times originally, trading off detailed
reporting for simplicity and intuitive parameter usage.

Change-Id: Ic653b9718640a2e5dc99ac819d7cb60dcc259f93
Reviewed-on: https://gerrit.franz.com:9080/13202
Reviewed-by: Yuri Zhykin <yuriz@franz.com>
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>