Handle input from stream for MAKE-HTTP-CLIENT-REQUEST
This change adds a STREAM-COMPUTED-CONTENT class which handles stream input
via chunking mechanism and implements implicit conversion from pathnames
and streams to FILE-COMPUTED-CONTENT and STREAM-COMPUTED-CONTENT instances
respectively. In order to automatically set required headers for custom
COMPUTED-CONTENT entities, the new GET-CONTENT-HEADERS generic function is
added. It takes the the protocol and headers provided to
MAKE-HTTP-CLIENT-REQUEST and returns new headers to be added to the set. It
is expected to signal errors in case of header/protocol conflicts.

Change-Id: I72898d10b917e29896046e39ce43cab0ab1d9379
Reviewed-on: https://gerrit.franz.com:9080/13772
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>