rfe16252: webactions 1.20: set http return code, convert src= links
spr44636 - instead of return a code of 200 allow the return code
   to be specified in the clp file  (using clp_response)

spr44527 - when processing a clp page convert links found in the
   src=xxxx parameter inside frame, script and img elements.

spr44597 - use a version of probe-file that will silently treat
   a file as non-present if the syntax of the filename is illegal.

Also in the makefile have aserve testing also do webactions testing.

%Testing: make test-aserve

<release-note>
Added the clp_response function to allow the clp file to specify the
HTTP return code when that file is returned to the browser.
The same conversion of links done to the value of href= is now
done to the value of src= in frame, script and img elements.
</release-note>
<documentation>
webactions.md updated.
</documentation>

Change-Id: I93a1c38b9d57487138f3df0b2db02194c8f2d81d
Reviewed-on: https://gerrit.franz.com:9080/13907
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>