rfe16267: add directory listing support
Directory listings are controlled via the :access-file argument to
publish-directory.  The syntax of access files has been extended to
allow:

  (:directory :list t [:hidden t/nil]
              [:files t/nil] [:directories t/nil])

This results in the contents of the directory in which it appears to
be listed when the appropriate URI is requested by a client.
Unauthorized files, given the above options or other access-file
options, are excluded.

Change-Id: I2f5f337dba25d40cdf527d28896bf9cbe2751108
Reviewed-on: https://gerrit.franz.com:9080/14037
Tested-by: Kevin Layer <layer@franz.com>
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>