rfe16407: put limit on Content-Length accepted
When get-request-body is called if the Content-Length is
greater than a specfied amount an error is signalled.
This will prevent an attack on the webserver causing it
to create a giant string object in memory.  This limit
does not apply to streaming the content, only to the
get-content-body call that wants to read all the content
at once.

A max-content-length slot was added to the wserver object
and it can be read or written with
the wserver-max-content-length accessor.
The default value is nil meaning there is no limit.
If you have a public facing server you should set a limit
based on the kind of data you expect to be sent to the server.

The net.aseve:start command takes a :max-content-length
argument which, if provided, will set that value in whatever
wsever object is used for this invocation of the server.

%Testing: cd aserve && make test

<release-note>
ref16407

New argument :max-content-length to net.aserve:start.
If a request comes in with a Content-Length greater
than the specified value an error is signalled
and no response is sent to the client.  This will
prevent a malicious client from causing the server to
run out of memory.

</release-note>
Change-Id: Iba4197ca6daec31a6a75c3b637859d9c7ce0bc65
Reviewed-on: https://gerrit.franz.com:9080/c/aserve/+/15850
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>