1.3.84: Add seize-request function
In order to handle requests that may take time for a
response the caller can now take charge of the request
from the aserver worker and place it, say,  in a table to be
handled later, thus not using up a thread for a long time.

The two new functions are seize-request and seize-request-finish.

%Testing: cd aserve; make test

<release-note>
1.3.84
Added a new function seize-request to allow user code to take
control of a request away from an aserve worker thread.
</release-note>
<documentation>
documentation (aserve.md) has been updated
</documentation>

Change-Id: I967413508229a155008c86c96c8da6f5ef140707
Reviewed-on: https://gerrit.franz.com:9080/c/aserve/+/20958
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>