1.3.85: support http on an ssl port
Add two arguments to net.aserve:start

http-on-ssl-port - if true then if an http requests comes in to a port
configured for SSL, then handle it as a http, not https, request. Without
this switch enabled the ssl layer in aserve will not understand the request
and it will fail

redirect-http-to-ssl - if true and if http-on-ssl-port is true than
an http request to a port configured for SSL will be redirected to
an https request to the same URL

Tests added.
Documentation updated.

%Testing: cd aserve ; make test

<release-note>
Optionally recognize and handle http requests made to an SSL port
</release-note>

Change-Id: I06caf207b42b659d02acc1a272ccbe13b49b286b
Reviewed-on: https://gerrit.franz.com:9080/c/aserve/+/21368
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>