1.3.88: bug27087: keep-alive and redirects
When do-http-reqeust is passed :keep-alive t it return as the fifth
value the socket connected to the remote server if the remote server
agreed to the KeepAlive request.  However if the server returned
a 302 redirect and then do-http-request followed the redirect and
retrieved that page then the socket returned to the client as
of the server to which the request was redirected.   This would
be confusing the client.
With this change  the fifth value will be nil (no socket) if
keep-alive was true and a redirect occured.  nil is always
a possible return value when :keep-alive is true as the
server may refuse to honor the KeepAlive request.

Test added.
No documentation update needed

%Testing: cd aserve ; make test

<release-note>
If keep-alive true is passed to do-http-request and
a redirect is done then the server connections will
be closed and nil returned as the fifth value.
</release-note>

Change-Id: I6d34e4017181ebf30c1af284ec542eae4ff6680f
Reviewed-on: https://gerrit.franz.com:9080/c/aserve/+/23490
Reviewed-by: John Foderaro <jkf@franz.com>
Tested-by: Kevin Layer <layer@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>