1.3.87: rfe17200: handle redirects before calling `read-body-hook`
Main client API function `net.aserve.client:do-http-request` handles
redirects unlike `net.aserve.client:make-http-client-request`, but
does not return client request objects, so for users who want to get
access to the client request object for some reason but with redirects
handled first, the only way is to use the `read-body-hook` parameter
of the `net.aserve.client:do-http-request`, which is a function that
receives the internal client request object as its argument. In order
for this to work, `read-body-hook` must be called *after* redirects
have been handled.

%Testing: cd aserve; make test

Change-Id: I8b8b38926dcfeaea1390ace145807b82fab82fd0
Reviewed-on: https://gerrit.franz.com:9080/c/aserve/+/23524
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>